% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_and_download_utils.R
\name{fetchIfMissing}
\alias{fetchIfMissing}
\title{Fetch File If Missing}
\usage{
fetchIfMissing(destfile, url, timestamp_file, cached_files, method = "curl")
}
\arguments{
\item{destfile}{Where to save file, if it's missing}

\item{url}{URL to use to fetch destfile}

\item{timestamp_file}{Optional file to record timestamp when new destfile is downloaded}

\item{cached_files}{Optional files that should be considered out-of-date and removed, if present when new destfile is fetched}

\item{method}{Method to use when downloading; passed to download.file.}
}
\description{
Fetch file if missing, in which case, record timestamp when
downloaded, and remove any specified cache files.
}
