% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_and_download_utils.R
\name{removeIfPresent}
\alias{removeIfPresent}
\title{Remove File If Present}
\usage{
removeIfPresent(filename)
}
\arguments{
\item{filename}{Name of file to remove, if present}
}
\description{
Remove a file if present.
}
